added guild_categories and database standards and converted all value types to reflect db

+++ internal/models/guild_messages.go
@@ -1,9 +1,9 @@
 package models
 
 type GuildMessages struct {
-	ID        int    `json:"id" db:"id"`
-	SenderID  int    `json:"sender_id" db:"sender_id"`
+	ID        uint64 `json:"id" db:"id"`
+	SenderID  uint64 `json:"sender_id" db:"sender_id"`
 	Message   string `json:"message" db:"message"`
-	ChannelID int    `json:"channel_id" db:"channel_id"`
-	SentAt    int    `json:"sent_at" db:"sent_at"`
+	ChannelID uint64 `json:"channel_id" db:"channel_id"`
+	SentAt    uint32 `json:"sent_at" db:"sent_at"`
 }
